+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bulk editing statements works again. (Closes: [#308](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/308))
 - Email addresses are now case-insensitive when logging in. (Closes: [#397](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/397))
 ### Added
+- Support formatted labels for all resources. (Closes: [#166](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/166))
 - Debugging database issues in tests is now described in the developer documentation.
 
 ## [0.15.0] - 2022-08-26